LPD-38105 adds set-up-tests-docker.gradle
This file is very similar to the local setup, but the server startup and client
extension deployment is handled by startDockerContainer instead of a bespoke
serverStart task.

The microservices are still started with a local background process. Manually
configuring the creation and starting of Docker containers is very difficult and
would be more easily managed by Docker Compose.